feat: add POST route and create error_handler
On this commit we added a temporary handler to the POST method, and we
added a new file, containing exception handler to all mapped errors.
