+++ meowpi/error_handler.py
@@ -1,0 +1,22 @@
+from fastapi import Request
+from fastapi.responses import JSONResponse
+from meowpi.utils.errors import UserAlreadyCreated, UserNotFoundError
+
+
+
+
+def setup_exception_handlers(app):
+
+    @app.exception_handler(UserNotFoundError)
+    async def not_found_exception_handler(request: Request, exc: UserNotFoundError):
+        return JSONResponse(
+            status_code=404,
+            content={"message": f"User not found"},
+        )
+
+    @app.exception_handler(UserAlreadyCreated)
+    async def alread_created_exc_handler(request: Request, exc: UserAlreadyCreated):
+        return JSONResponse(
+            status_code=400,
+            content={"message": f"Username is already being used"}
+        )
