+++ meowpi/main.py
@@ -1,18 +1,11 @@
-from fastapi import FastAPI, Request
-from fastapi.responses import JSONResponse
+from fastapi import FastAPI
 from meowpi.database.mongo import mongo_database
 from meowpi.models.users import CreateUserSchema
-from meowpi.utils.errors import UserNotFoundError
-
+from meowpi.error_handler import setup_exception_handlers
 
 app = FastAPI()
 
-@app.exception_handler(UserNotFoundError)
-async def not_found_exception_handler(request: Request, exc: UserNotFoundError):
-    return JSONResponse(
-        status_code=404,
-        content={"message": f"User not found"},
-    )
+
 
 @app.get('/')
 async def meow_world():
@@ -25,4 +18,7 @@ async def get_user(username: str):
 
 @app.post('/user')
 async def create_user(create_user_data: CreateUserSchema):
-    return "aoba"
+    user_id = await mongo_database.create_user(create_user_data)
+    return user_id
+
+setup_exception_handlers(app)
